requirements.txt
django-crispy-forms==2.0
django-filter==23.2
djangorestframework==3.14.0
crispy-bootstrap4==2022.1